client_agent/requirements.txt
websocket-client==1.9.0
mss==10.2.0
pillow==12.2.0
requests==2.33.1
psutil==6.1.1
